Converted project to compose
* Displaying Movie List

- Added Composables to display a Lazy Grid of Movies
- Removed unneeded classes

* Compose Move List Working

Heavy clean up of unneeded classes.

* Redoing of ViewState and Navigation

* Movie List and Movie Details have their own top bar

Also removed defect with bottom nav showing up on details screen

* Movie Details cleaned up

Top bar, backdrop, title, poster and details in place

* Previews and Reviews Step 1

They render!

* Removing Unneeded Files

* Adding Color And Theming

* Correcting Colors and Cleanup

* Adding expandable review text

* Can add and remove favorites

* Completely use lazy column for details screen

* Created an App State to handle bottom nav along with navigation refactor

* Show/Hide Reviews/Previews Sections if needed.

Further padding tweaks done.

* Landscape mode tweaks

+++ app/src/main/java/org/sco/movieratings/moviedetails/MovieDetailsViewModel.kt
@@ -1,11 +1,11 @@
 package org.sco.movieratings.moviedetails
 
-import androidx.lifecycle.*
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.viewModelScope
 import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
+import kotlinx.coroutines.flow.*
 import kotlinx.coroutines.launch
-import org.sco.movieratings.api.response.Preview
+import org.sco.movieratings.api.response.MoviePreview
 import org.sco.movieratings.api.response.Review
 import org.sco.movieratings.db.MovieSchema
 import org.sco.movieratings.repository.MovieRepository
@@ -16,36 +16,33 @@ class MovieDetailsViewModel @Inject constructor(
     private val movieRepository: MovieRepository
 ) : ViewModel() {
 
-    fun getReviews(movieId: Int): LiveData<List<Review>> = liveData {
-        emitSource(movieRepository.getMovieReviews(movieId).asLiveData())
+    fun getReviews(movieId: Int): Flow<List<Review>> =
+        movieRepository.getMovieReviews(movieId)
+
+    fun getPreviews(movieId: Int): Flow<List<MoviePreview>> =
+        movieRepository.getMoviePreviews(movieId)
+
+    private val _isFavorite = MutableStateFlow(false)
+    val isFavorite: StateFlow<Boolean> = _isFavorite
+
+    fun checkIsFavorite(movieId: Int) = viewModelScope.launch {
+        _isFavorite.value = getMovie(movieId).firstOrNull()?.let {
+            movieRepository.isFavorite(it)
+        } ?: false
     }
 
-    fun getPreviews(movieId: Int): LiveData<List<Preview>> = liveData {
-        emitSource(movieRepository.getMoviePreviews(movieId).asLiveData())
-    }
+    fun getMovie(movieId: Int): Flow<MovieSchema?> =
+        movieRepository.getMovie(movieId)
 
-    private val _isFavorite: MutableStateFlow<Boolean> = MutableStateFlow(false)
-    val isFavorite : StateFlow<Boolean> = _isFavorite
-
-    fun checkIfFavorite(movieSchema: MovieSchema): LiveData<Boolean> = liveData {
-        movieRepository.isFavorite(movieSchema).collect {
-            emit(it)
-            _isFavorite.value = it
-        }
-    }
-
-    fun addToFavorite(movieSchema: MovieSchema) {
+    fun onFavoriteClick(movieSchema: MovieSchema) {
         viewModelScope.launch {
-            movieRepository.addToFavorites(movieSchema)
-            _isFavorite.value = true
+            val favorite = movieRepository.isFavorite(movieSchema)
+            if(favorite) {
+                movieRepository.removeFromFavorites(movieSchema)
+            } else {
+                movieRepository.addToFavorites(movieSchema)
+            }
+            _isFavorite.emit(!favorite)
         }
     }
-
-    fun removeFromFavorites(movieSchema: MovieSchema) {
-        viewModelScope.launch {
-            movieRepository.removeFromFavorites(movieSchema)
-            _isFavorite.value = false
-        }
-    }
-
 }
